Add a media settings view for images (#1527) Ref: http://ios.trac.wordpress.org/ticket/1527
Added a media settings view that allows a user to manage the caption, link, alignment, position and size of an image that has been uploaded from the iOS device. Changes made to the media settings on other devices (web, android, etc) are handled. Example, you can upload an image from your phone, edit the size and caption on the web site and then open the post on the phone and make changes to the settings again. If there are any additional attributes added to the image or anchor elements they will not be lost.

Note: Images uploaded from other devices are still not displayed in the attachments for a post.

* Updating media no longer blindly replaces text, handles when the user
makes changes to the HTML manually or from another device, and keeps
attributes in place that the media settings can not handle
* Changed all the media notifications to be sent in the same way. The
notification object is always the sender and Media and MediaSettings
are sent inside the userInfo dctionary.
* MediaSettings
** Class method createMediaSettingsForUrl can be used to pull out the
HTML for a media object matching the URL. This HTML will include the
caption, a and img elements.
** 5 properties are exposed that can be set from the new
MediaSettingsViewController (captionText, alignment, customWidth,
customHeight, linkHref). Making changes to any of these will only
replace/update the text in the media HTML and leave all other
attributes alone.