Quick post performance, progress and other minor improvements
* Improve the performance of uploading images.
** The compression of the JPEG images has been increased. The quality
was originally set to 90% but after some testing and research it seems
that 60% should be more than acceptable. This should reduce the file
size by ~50% for the images being uploaded.
** In the quick post editor as soon as an image is selected it starts
uploading so while the user is writing the title/content the image can
hopefully finish. Once the user taps the Publish button only the post
has to be uploaded. If the user selects a different blog the image is
removed from the original blog and then queued to upload to the newly
selected blog.
* Add a circular progress view while the quick post image is uploading.
Once the image has completed the icon will switch to a spinner while
the post is uploaded.
* Add an action sheet to the quick post button while the post/image are
being uploaded to allow the user to stop or remove the post being
uploaded
* Add an action sheet to the quick post button to allow the user to
easily open the previous failed quick post inside the quick post editor

* Bug fix for #1472 - Set the quick post to uploading right away so
that the user can not edit the post while it is being uploaded
* Fix the blog selector button so the 'g' is not cut off
* Fix to hide the paperclip when the quick post picture is zoomed
* Add a block of code for opening a post in the full editor for the
quick post editor to use (unused right now)