Fixed a crash that occurs when removing a blog from the settings screen. h/t @jkv
The crash would occur if you would attempt to sign into a self hosted blog through the new
sign in screen and forgot to put in the website initially. After signing in if the user opened up
the settings page then removed the blog, the app wouldn't sign out and opening up the Sidebar
would result in a crash.