Improved cookie verification
Fixes #1593 and who knows what else.

Previously, we checked for the existence of a wordpress_logged_in cookie
to decide wether we needed to authenticate or not. We didn't check if it
was for the right user, so best case scenario the user would end up with
a login page.

Reader checks were similar, so even after sign out and sign in with a
different account you got the previous user posts in the reader, at
least until it refreshed.

On top of that, the generic web view was checking for cookies on
kMobileReaderURL instead of the URL we were trying to load.