Modified notification detail to be more secure when opening sites.
Fixes: http://ios.trac.wordpress.org/ticket/1642

If a user tapped on notification detail views for follows and likes it would result in a WPWebView being loaded up with the user's wordpress.com credentials so as to allow for the viewing of private blogs. The problem with this was that this was done under http and not https and there was no check performed to see if the site in question was a wordpress.com site. This same issue happened if the user clicked on the header portion of a notification detail view. This fix ensures that credentials are only sent to wordpress.com sites and that https is used when sending credentials to wordpress.com