Fixes missing refresh header on notifications view.
There was a xib for the notifications list view. Redundant since its inheriting from WPTableViewController.  The loaded xib resulted in two UITableViewControllers, one a child of the other.
Removing the xib and calling init instead of initWithNibName:bundle fixes the issue where the refresh header would sometimes appear to be missing when scrolling the extra UITableView.

Also removes a now obsolete xib for the hybrid reader.