Fixes an issue where quick post/page + buttons showed the wrong state.
The UITableViewCell was setting the accessory view's highlighted state to YES. Not sure if this is happening in the code somewhere and I couldn't find it or if this is something Apple does by design. Setting the highlighted state to NO when the cell's selection state changes seems to fix the problem.