Fixes layout concerns with the View Admin button.
The view admin button would overlap the view site button, and show up even when the section was closed.
Restores the View Admin cell and populates it with a button. The button intercepts the tap on the row and avoids triggering row selection.
Seems to work well enough tho the button icon dims when pressed and the icons in cells do not.  Minor difference.