Decide to load images depending on scrolling speed
Before this, we loaded reader images only when the table wasn't
scrolling or being dragged unless we had the image ready with the
correct size. This meant that images wouldn't load if you were scrolling
really slow, or even stopped but still had your finger on the screen.

This implementation only prevents image loading if we detect the table
view is scrolling "fast".

After testing, images seem to load almost immediately on WiFi while
performance is still good. I'll have to test on 3G to see if it's good
enough.

see #24