Fixes an issue where simulating pull to refresh tripped isFastScrolling
Setting the table view's content offset, as we do when simulating pull to refresh or dismissing the refresh header, triggers the scrollViewDidScroll: delegate method.
Since the refresh offset is greater than the fast scrolling threshhold the _isFastScrolling flag is set but never cleared since the scrollview is stationary.
The result is that newly added cells do not display their featured images.
Adds a flag that we can check to see if the refresh offset is being set so we can skip setting the _isFastScrolling flag