Fix crash when adding blogs
When the user only had one blog and it's already added, trying to add
more wp.com blogs might result in a crash.

`refreshBlogs` was called on `viewWillAppear`, and it might end before
the view is fully presented. This results in popping the view controller
while it's still being animated and crashes.

Call `refreshBlogs` from `viewDidAppear` instead so the animation is
completed.