Added loading image behind the signed out state for the iPad.
Fixed an issue where if the user signed out on an iPad they would get
a form sheet with the NUX but this form sheet didn't hide the sidebar
related views behind it so it looked rather ugly. To prevent this we
place the loading image on top of the sidebar related views.