Removing unnecessary code that manually set back button in editor.
We put these lines of code in to deal with some situations in iOS6
when a back button wouldn't appear. The reason the back button
wouldn't appear was because we removed the title from the editor. As
we've restored the title in 9bf631c949a7cfe724fbdb3819fc75336188812b
this code is no longer needed.