Fixed localization issue when displaying 'no results' text.
The previous 'no results' text was "No %@ yet" with the localized name
of the item being swapped into the above string. The problem with this
is that it doesn't localize well. As such we swapped this string out
with individual strings for each type so as to enable better
localization. Fixes #214