#457 - Unregister from push notifications before removing WP.com credentials
Unregister call to WP.com requires username and password to unregister a push token.  The app used to send NSArray arrayWithObjects and the null password would terminate the array, not causing a crash.  The unregister command would never work, however, and fail silently.  The recent change to how the array is created brought this to light - now unregistering for push notifications before the actual acccount is deleted from the app.