Set post failed directly insead of using delay
The logic behind the old code was that, according to the docs, changes
to the object were ignored during `awakeFromFetch`.

That's not accurate, what the doc says is:

> The managed object context’s change processing is explicitly disabled
> around this method so that you can use public setters to establish
> transient values and other caches without dirtying the object or its
> context.

From what I've read this means that the generated dynamic accessors
won't change the actual values, but you can still use
setPrimitiveValue:forKey: to change the object.

I don't think this will actually fix #623, but it might.

Using `performSelector:` did schedule `markRemoteStatusFailed` on the
next run loop cycle. For reasons that completely escape me, this was
being scheduled in Quantcast's library internal run loop, which was
obviously not running on the main thread.

I don't think that was the main cause of the crash (could not fulfill a
fault), but I can't imagine it did any good.

If I understand this right, and nothing else fixed it, at least this
will crash on `setPrimitiveValue:forKey:` and not on the delayed call,
providing us with the exact backtrace of whatever is calling this on a
secondary thread.