#673 - Added OAuth2 token to API call to verify WP.com connectivity
Updated the app delegate to include the OAuth2 token in the call out to WP.com for verification of credentials.  This is NOT the ultimate solution.  Auth verification should be a function of WPAccount.  Removed the NSUserDefaults for the flag to indicate if the app is authenticated for WP.com - presence of OAuth2 token is the best guess at auth status upon app launch.

Updated minimum WordPress install version to 3.5 per our release notes.

WP.com-centric UI does not handle a revoked OAuth2 token properly (Reader for example)