Always show the post tab, even in the rare case there are no blogs.
When no blogs and the post tab is tapped present the login view controller to the user.
After logging in and selecting a blog show the editor.
Moves the default blog logic out of the Blog model and keeps it in the EditPostViewController.
Saves the last edited blog's url and uses that blog for new posts.
If no saved blog default to the first one returned.