Use WPOAuthClient for sign in
Note @koke, that for JetpackSettingsVC, the effect of creating a new account may not be the correct approach taken. signInWithUsername:password:blocks use to store the authtoken internally within WordPressComApi. Now, the only way to save a token is by creating a new account.

We could make `refreshTokenWithSuccess:failure` useful again for these kinds of purposes.