Setting the locale in the DateUtils to English POSIX to prevent conversion failures
Using date strings in the date formatter is tricky - when interpreting dates from XML (or wherever) the formatter has to be locked to a locale so the letters used mean the right thing.  Device locale can affect what the letters mean - in this case setting to 12hr en_GB fails when converting a date over the wire.