Cleans up branching in PostSettingsViewController
Cleans up some hard to read branching in the PostSettingsViewController. More cleanup to follow.
Stubs out new cells to encapsulate featuredImage and geoloction functionality
Removes some unused views from the PostSettingsViewController xib.

This commit breaks the VC for pages. We'll restore pages functionality in a future commit.
This commit breaks the row height sizing for geolocation and featured images. We'll restore this in a future commit.