Reload table when blog list changes
The blog list will not change often, and trying to manage rows/sections
individually is tricky, specially when there’s not a 1-1 match to the
data source (“Add site” row).

Reloading the whole table when the results change is less prone to
errors and shouldn’t affect performance much, since it doesn’t happen
often.