Only consider self hosted for the welcome screen
We are already checking for an account to see if we are logged into
WordPress.com.

The reason for this is that now the account changed notification is
posted as soon as the default account is set to nil, but at that point
the account and its blogs still exist.

The `noBlogsAndNoWordPressDotComAccount` method would always return `NO`
on sign out, since deletion is done asynchronously.