Changed tapping the 'Me' tab behavior for users with a single site.
Modified the tapping of the 'Me' tab to bypass the main blog selection
screen and just show the individual blog details as the vast majority
of our users will only have one blog and adding an extra click is a
bit redundant for them. If the user taps the 'Me' tab while the tab
is currently active it'll pull up the blog list, but if they select
the 'Me' tab while the tab isn't active it'll pull up their blog's
detail screen.