Fix future date issue on months for stats.
If the user clicked on the months filter of the visitors and views
section it could appear like the final month was a future date as the
months would display like "Mar 14" with just two digits on the year.
If the year digits were greater than the current date it could
possibly confuse the user. To fix this I just adjusted the label to
print out four digits for the year.