Modified "Need help?" link for 2fa errors to open correct page.
Should the user enter their WordPress.com password and get back the
error message that they need to generate an application specific
password, the need help? link used to open the support center but
now with this change it should open up the page on wordpress.com that
tells a user how to generate an application specific password.