Fake success if a duplicate comment fails
The most common scenario for this error is bad network conditions: we
tried to post, the comment reached the server but we never got a reply.

When the user tries to post again, the server replies saying it already
has that comment.

Since that's what we originally wanted, we treat it as success

Fixes #498