Split jetpack auth/skip funnel into two steps.
Previously we had two events, "Add Self Hosted Site and Connected to Jetpack"
and "Add Self Hosted Site and Skipped Connecting to Jetpack". We
decided to split this up into three steps:
1. Added Self Hosted Site
2. Skipped Connecting to Jetpack
3. Connected to Jetpack

The reason for splitting this up into more steps is that it makes it
easier to synchronize funnels with the Android app ad Android doesn't
prompt the user to connect to Jetpack right away, only when they want
to use the stats.