Added REST support for blog syncing
Note that the syncBlogsForAccount: method is only called on the default
WordPress.com account, so we might as well drop XML-RPC support for it.

Refs #920 #1521

*Warning:* Jetpack sites are returned by me/sites but they won't work.
There's a bug in the API that returns the xmlrpc URLs as HTTPS, and we
don't even have the self-hosted credentials. We'll need to filter those
out until we fully support REST.