Preserve orientation on optimized images
A commit message might not be the best place for a rant, so go read this
http://www.daveperrett.com/articles/2012/07/28/exif-orientation-handling-is-a-ghetto/

In essence, yet another chapter of "there's a standard for this, but
everyone's doing it differently".

Anyway, it seems the safer choice is to interpret the Orientation data,
rotate the image, and strip the metadata.

Otherwise, episodes of black magic might occur. Like Chrome on OSX
rotating different than Chrome on iOS

https://cloudup.com/cbaGIbkscyO

Or Chrome rotating different than Chrome's inspector

https://cloudup.com/cJfi_bdJiKR

Using the test images from Dave Perrett's post,

Before: http://breakmeplease.wordpress.com/2014/05/28/exif-test-for-1685/
After: http://breakmeplease.wordpress.com/2014/05/28/exif-test-for-1685-remove-metadata-pass-orientation-use-correct-size/