Updates for removing/updating followed/unfollowed posts.
Adds a way to update the following status for all posts belonging to a particular site.
Adds a way to delete posts belonging to a specific site from a specific topic
Updates the post merge cleanup logic to use the startDate of a batch of posts vs the date of the earliest post. This catches some stale posts that would other wise remain.