Add semi-transparent header subview
Changed NotificationDetailsViewController subview adding order to
ensure replytextview is added last.  Changed SuggestionsTableView to
inherit from UIView and now it includes a semi-transparent header view
and a simplified table height constraint calculation.  This way when
there are fewer results than will fit, the semi-transparent header will
appear above the results.  It will also be easier to modify the
SuggestionsTableView now to work with posts (where no transparent
header is required, but a footer will be.)