Drop batched operations
The way we use callbacks for our service and remote methods makes it
hard to batch operations when an error occurs.

If we're batching two operations (say option and formats sync) and one
fails, what do you do?

- Call `failure` once?
- Call `failure` twice?
- Do you call `overalSuccess` at all?

Also, the REST `/batch` method is not well documented

To keep it simple, the remote blog services now support single requests
only. If we move to a notification based syncing instead of callback
blocks, batching could become an interesting option