Prevent double prompt of app reviews in short span
Right now the app is set to prompt the user upon a certain set of conditions after a new release. As we are hot fixing 4.6 the day it was released this will result in a few users receiving double prompts in a short time span which is undesirable. As such we decided to disable the prompt for an app review of 4.6.1 if the user had interacted with the app review tool in any form.