Removed warning for unknown keys
Removed the warning that was spit out by `gencredentials.rb` if it didn't know what to do with a key. Some people remove keys by renaming them or have extra keys for other things they are testing so it's annoying to see warnings appear all the time that aren't really needed.