HelpshiftUtils introduced to abstract Helpshift logic in different places.
A new setup method added to initiate all the setup required for Helpshift. Instead of changing the Helpshift delegate in many different places HelpshiftUtils singleton will listen for the changes and broadcast them to other view controllers. WPTabBarController should update the badge count for the 'Me' tab itself. If the current view is going to show the unread count for it will request an update on it, so we keep the unread count fresh all the time. A new badge is added to the 'Me' page 'Help & Support' row to show the unread count.