iPad logout issue resolved
In iPad we are showing a `UIActionSheet` for the Sign-out button and it acts as a `presentedViewController` for the `WPTabBarController`. Since we check `presentedViewController` and dismiss it in the `AppDelegate` before the login screen appears, that UIActionSheet was preventing login page from showing up. This case is now ignored and login page should appear even if the action sheet is not dismissed yet.