Do not reusue label for cell accessory view.
The freeze was caused when two cells were fighting over who had ownership of the label used for the accessoryview.
This created an infinite series of calls to refresh view heirachy between the two cells as one gained the view
and the other lost it, triggering it to take it back, etc.