Fix Jetpack preview authentication
As a side effect of #3597, `blog.authToken` was not `nil` for Jetpack
sites. When there's an auth token, WPURLRequest will drop the password
from the request. Since the login was posted to the self hosted and not
WordPress.com, the authentication failed.

This sets the token only if the blog is a WordPress.com one, and adds a
safeguard to prevent tokens from being sent to non-wp.com URLs.

fixes #3637