Fixed web view rendering
On #2679, a hack was introduced to set the viewport width to the actual
web view width.

The problem is some sites set the viewport width to `device-width`,
which is a problem if the web view is in a modal window instead of full
screen.

This change only applies the hack if the web view's width is less than
the window's width