Adjusting fabric run script to use a separate key
Fabric technically requires a slightly different key than crashlytics as there's a second parameter that goes with it. On some of our computers it seems to work ok with just the first key but on others it doesn't seem to work without the second one. Lets create a separate entry for the key needed by the script just to be safe.