Make primary blog first
Sort the site list and site picker so that the primary blog appears
first.

I'm not too happy with the readability of the code but it seemed like
the simplest option: since core data relationships have to have an
inverse, `accountForDefaultBlog` will only be set for the default
account's default blog. It will be nil for all the other blogs

Fixes #2599