Disable incomplete umbrella warnings
After the migration to frameworks we're having some trouble with Release
builds.

What I've understood so far:

WordPress depends on Stats and the Today widget depends on
Stats/Services but either cocoapods or xcode can’t handle having a
duplicate framework dependency.

In
`DerivedData/WordPress-awgiklmsmdriyvfpppixcrczellm/Build/Intermediates/ArchiveIntermediates/WordPress Internal/IntermediateBuildFilesPath/UninstalledProducts/iphoneos/WordPressComStatsiOS.framework`
I have a modulemap that includes the umbrella header
"Pods-WordPress-WordPressCom-Stats-iOS-umbrella.h"
but inside Headers:

```
ls Headers/*umbrella*
Headers/Pods-WordPress-WordPressCom-Stats-iOS-umbrella.h
Headers/Pods-WordPressTodayWidget-WordPressCom-Stats-iOS-umbrella.h
```

So it looks like it generates the headers and module for one of the
targets then it overwrites the framework and modulemap with the second
target but the previous umbrella header is still there and it’s not
included in the modulemap

Adding `-Wno-incomplete-umbrella` fixes the error until we can figure
out how to avoid the duplicated framework problem.

Pushing directly as @sendhil reviewed on Slack