Fixes section headers on Me
I merged #4589 a bit too fast, and the section header title for the
WordPress.com account sign in/out looks out of place on the iPad.

While we find a global fix in #4591, I've skipped ImmuTableViewHandler and
implemented the data source and delegate methods on MeViewController, so I can
provide a WPTableViewSectionHeaderFooterView as the section header view.