Reader: Adds a nil check before creating properties dictionary.
Unsure of the situation but there are at least some cases were either the postID or the siteID (or both) can be nil when the success block is called.
Assign the postID and siteID to NSNumbers and let them be retained individually by the block. Then perform a nil check.