Blog Details: Quick refactor making it easier to change table content.
The forthcoming sharing feature will be enabled/disabled via a feature flag and accesible from the Blog Details screen.
This change makes it a bit easier (and saner) to present cells conditionally, and avoids some complicated const wrangling.
The approach is similar to what was done in ImmuTable and should be considered as a stopgap/stepping stone til ImmuTable is usable in ObjC classes, or when BlogDetailsViewController is refactored to Swift.