Refactor My Profile to improve memory management
Prior to this, MyProfileController was a class and owned the view controller.
This had a couple problems:

 1. We want to release the whole thing when UIKit decides it should deallocate
the view controller, and for that we need the VC to own the controller.
 2. Having the controller being a class led to a lot of retain cycles since the
code relies on closures a lot.

With the new code, the VC and controller are deallocated when the VC is popped,
and the Observable subscriptions are properly disposed.