Refactor ImmuTablePresenter/Controller/VC
The main goal is to remove `visible` from `ImmuTablePresenter`, since it
doesn't really belong there. Presenter is a thing that can "present" view
controllers: usually a UIViewController, but implemented as a protocol so it's
testable.

But then I realized, the Controller doesn't really need a reference to
Presenter, or the parent View Controller. The only thing that needs that is
ImmuTable (for the callbacks), and we can pass the presenter when requesting
the table view model (renamed). This is similar to the previous solution, but
avoids the chicken-and-egg problem when initiailizing. And because of this, it
removes 4 assertions/preconditions that are now guaranteed at compile time.

The `visible` observable is still exposed on ImmuTableViewController, but
the VC is the one pausing the subscription to tableViewModel and errorMessage,
instead of relying on the controller. I like this slightly less semantically,
but it makes everything much simpler with initialization.