Improve accessibility of plan details.
Before, VoiceOver would read all the features without saying if they were
included in the plan or not.

With this change, it will add either "Included", "Included in web version", or
"Not included" after reading the feature name. For features with a description
(support and storage), it will still read the name and description.

It adds some refactoring so the accessoryView and spoken text follow the same
logic.