Pointing stats pod to workaround branch for 6.0 because of broken app submission
App submission was broken because there was Swift in the Stats Pod that was referenced in the today widget which the app store didn't really like. I created a temporary branch that converted the small Swift file to Objective-C. Eventually I'll move this code upstream in the Stats pod and we can go back to pointing at specific version numbers for the stats pod again.