Alternate approach to PromptViewController
Implemented some of the suggestions in #5147 for comparison.

Ironically I had to move the `Comparable` protocol to Objective-C so it could
be enforced in Swift. When I started adding the generic constraints to enforce
that the presented view controller was both a UIViewController and a
Confirmable, Swift wasn't picking up that SettingsTextViewController was a
Confirmable, because it was conforming to the protocol in the implementation
file. And the conformance couldn't be moved to the header file because we can't
import WordPress-Swift.h from there, and we can't do a forward declaration of a
protocol for conformance.

Until all the use cases are written in swift, we still have to internally store
a UIViewController and cast accordingly, but now the public interface only
allows for a view controller that implements Confirmable, enforced by the
compiler when using Swift, and at run time via an assertion when used from
Objective-C.

I've also wrapped PromptViewController inside it's own navigation controller
and added a helper `prompt` method to immutable that enforces the right type.

I haven't addressed other suggestions, and this would still need better
documentation.