Prevent race condition when changing settings
Previously, when the SettingsTextViewController is dismissed, the value changes.
The service applies the change locally and saves immediately, then queues a
change request. When Profile/Settings appears again it triggers a remote
refresh, which brings back the old value. When the change request completes, it
doesn't update the local value.

The downside of this fix is that the VC will keep getting updates even if not
visible but I can live with that for now and the alternatives are much more
complicated.