Posts/Pages: 999 priority on trailing margin constraints.
This fixes an issue seen when transitioning from full-screen to compact
width multitasking states, on iPad. In this case, trailing margin
constraints break, or at least complain in the console. Changing the
priority to 999 silences the breakage in the console, but it seems
there is still an iOS layout transition going on here for multitasking,
although the constraints still resolve correctly in the end.