Pass the stream topic to the reader details
Some of the details logic was based on the currently viewed topic. This
information was previously not passed from the stream, but inferred from
the post's topic. Now that the topic-post relationship is many to many,
this assumption is not correct. Instead by explicitly passing the topic
to the details we can be sure we are enabling the correct logic.