Reader Comments: Reloads tableview when view appears but not when app becomes active.
There is a conflict between reloading the tableview when the app becomes active and state restoration.
In such a situation, calls to tableView.beginUpdates/tableView.endUpdates collide with an
already inflight CATransation and output warnings to the console. Additionally, cells containing
image attachments can end up with the wrong height, or cells made obsolete due to other cells resizing
can fail to be removed from the tableview and potentially overlay cells in use.

Reloading from viewDidAppear instead avoids the previously mentioned conflict, and
ensures that cells are properly sized/resized.