Don't show errors if load more fails on posts.
Just as we don't show errors on sync unless it was user initiated, avoid
presenting an alert if infinite scroll was triggered.

The deeper issue is that the infinite scroll code is triggered from
willDisplayCell which doesn't only get called as you scroll, but also whenever
we update the table data (i.e. every core data save).

There is a lot of room for improvement there, but this will at least avoid
presenting the user with countless alerts when trying to write a post while
being offline.