Force Left-To-Right for NUX fields.
All those fields (username, email, URL) enforce ASCII characters. I'm not sure
about password, but it looks like the web version doesn't flip those on the sign
up flow.

wp-login seems to have the form fields flipped.

For us, when creating a new site, the `.wordpress.com` suffix on the URL field
poses a problem as it can't really be flipped and look right, so LTR makes more
sense in that view. And it's probably best to keep it consistent for all of
NUX, so LTR it is for everything.

We can definitely revisit that decision, but at least for now, forced LTR is
still probably better than broken RTL.