Replace fix-translation script with a Swift one.
Trying to make sense of all the different encodings in PHP was too painful.
Swift makes this so much easier by abstracting out all the painful details of
dealing with UTF-16.

I've dropped the part that was escaping quotes as it seems we don't have any of
those now, and if they came back it should be handled on GlotPress side.