Mark notifications read only when they're displayed
This switches the responsibility for marking a note as read to the detail view
controller, since it'll be able to tell if it's visible or not.

Previously it was marked as read when the view was loaded, regardless of
visibility: split view would preload the first notification even on iPhone or
if the user was in different tab.