Improve scrolling performance for edit sites
Instead of fetching the default account every time we check for visibility
support, we just need to check if the account's uuid matches the stored one.

We could optimize further, but I think the core data fetch was the big
bottleneck and I'm seeing a similar scrolling performance than the non-edit
list with this.