Improve marking notifications read logic
This handles several issues with the previous implementation:

1. NotificationDetailsViewController was listening for changes on its
notification, and marking it as read, even if it wasn't visible (app in the
background, split view keeping a reference to a detail even after dismissing
it). Now it listens to background/foreground notifications and only updates if
the view is visible.
2. NotificationDetailsViewController was marking notes as read from the
NSManagedObjectContextObjectsDidChange notification, causing a core data save
within another core data save, and crashing. Now it uses an async dispatch to
break out of the current save.
3. If the markAsRead request failed, it'll revert the note to unread, and
NotificationDetailsViewController would attempt to mark it as read again. This
was causing the original crash, but when the crash was avoided it was causing a
request loop and wasting CPU cycles (and battery). Now we'll pretend it
succeeded but invalidate the cache, so the next successful sync can mark it
unread again.