Fix site icon for Jetpack sites
This stops relying on blavatar for self hosted sites and will use the blog.icon
passed through photon.

To avoid some problems with resizing, the site icon size is fixed at 40x40 for
download. We previously tried to calculate it based on the imageView size, but
for some reason it's 74x73.5 by the time the cell is being configured